Started work on level5. Added new mapObjects and titlescreen
Added level5 and began work on it, themed around buttons. Will probably
move it from the 5th slot to later in the game.
Added new mapObjects, two Portals which send you to the next/previous
mapObject in the mapData array (meant to be used together), a button to
remove the previous mapObject, a button to change the previous
mapObjecct (normally used to add one), a pressed button (for when a
button has been pressed to change into a empty button) and a empty block
(to be used with changing button to add a new block (really just makes
block turn visable/hitable)).
Added new titlescreen with new artwork and bouncing O
Also added end of level effect where the level "loads out" when player
hits end marker by reducing the currentlLevelArraySize every 5 frames
Changed a few shorts to bytes where applicable