Added level between 7 and 8, cleaned up level 8(9 now)
Added a new level between 7 and 8 as a tutorial to bouncing up pipes,
and removing blockages with buttons. Needed to go before level 9 as
there is a non obvious pipe in level 9.
Cleaned up level 8 (which is level 9 now) by resizing some blocks to be
the correct size.
Added an execption to map celing kill trigger for level 8 (the new one)
because the level is very tall. Need to come up with a more elegant
solution in the future.