Added death feedback, delete save file option, made checkpoint hitboxes larger
Added player feedback on death, such as a sound when falling off the map
and a frame delay+sound when hitting a spike.
Added a save file delete option, which replaces the current save file
with a fresh one.
Increased the size of the checkpoint hitbox as it was possible to pass
through them.
Fixed a map's spike placement issue.
Fixed level select menu using wrong variable for storing selection.